광고 신청 기능 구현 중 + 전문가 등급 기능 구성 완료

+++ lib/expert/ad_model.dart
@@ -2,7 +2,7 @@ class Ad {
   final String id;
   final String title;
   final String description;
-  final double budget;
+  final int budget;
   final DateTime startDate;
   final DateTime endDate;
   final String userId; // 사용자 정보 추가
